+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Release date: UNRELEASED
 
 ### New features and improvements
 
-* ...
+* Added a `--orientation` option to the `pagerotate` command to conditionally rotate the page to a target orientation (thanks to @gatesphere) (#705)
 
 ### Bug fixes